refactor MLflow integration in RecordLogger and TD3; improve artifact logging conditions and add TrainingThread management
