+++ reinforceui_studio/RL_loops/training_policy_loop.py
@@ -88,8 +88,13 @@ def training_loop(  # noqa: C901
         config_data, render_mode="rgb_array", evaluation_env=True
     )
 
+    # todo find a better way to handle the experiment name
+    print(log_folder_path)
+    input("Press Enter to continue...")
+
+
     mlflow_logger = MLflowLogger(
-        experiment_name=f"RL_{algorithm_name}",
+        experiment_name= "experiment_name",
         run_name=display_name,
         tags={
             "environment": config_data.get("selected_environment"),
